pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lead-hunter"
version = "0.1.0"
description = "Automated B2B lead generation pipeline powered by CrewAI"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.25.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "PyGithub>=2.1.0",
]

[project.scripts]
lead-hunter = "lead_hunter.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src/lead_hunter"]

[tool.crewai]
type = "crew"
